Safari-609.1.11.1/LayoutTests/accessibility/mac/details-summary-role-description.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<details id="details" open=""><summary id="summary">Summary</summary><p>some detail</p></details>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that summary and detail elements have correct role description.");

    if (window.accessibilityController) {

          var summary = accessibilityController.accessibleElementById("summary");
          var details = accessibilityController.accessibleElementById("details");
         
          shouldBe("summary.role", "'AXRole: AXButton'");
          shouldBe("details.role", "'AXRole: AXGroup'");
          shouldBe("summary.subrole", "'AXSubrole: AXSummary'");
          shouldBe("details.subrole", "'AXSubrole: AXDetails'");
          shouldBe("summary.roleDescription", "'AXRoleDescription: summary'");
          shouldBe("details.roleDescription", "'AXRoleDescription: details'");
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
